requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
pymilvus>=2.4.0
requests>=2.31.0
sentence-transformers>=2.2.0
yt-dlp
